+++ test/animated_emoji_test.dart
@@ -1,9 +1,21 @@
+import 'package:animated_emoji/animated_emoji.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:animated_emoji/animated_emoji.dart';
-
 void main() {
-  test('adds one to input values', () {
-    print('HIO');
+  testWidgets('Can use widget', (tester) async {
+    await tester.pumpWidget(const AnimatedEmoji(AnimatedEmojis.angry));
+    expect(find.byType(AnimatedEmoji), findsOneWidget);
+  });
+
+  testWidgets('Has error Widget', (tester) async {
+    await tester.pumpWidget(
+      const AnimatedEmoji(
+        AnimatedEmojiData('Oh no'),
+        errorWidget: Text('error'),
+        size: 128,
+      ),
+    );
+    expect(find.text('error'), findsOneWidget);
   });
 }