+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## 1.0.0
+## 1.0.1+1
+
+* **deprecated**:  `simpson` skintone. Use a value of `null` instead.
+* **doc**: Updated readme to use WEBP.
+* **refactor**: now using [very good analysis](https://pub.dev/packages/very_good_analysis)
+* **updated**: tests
+
+# 1.0.0
 
 * initial release.